+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 - Update OpenTelemetry dependencies to 0.18.0
 - Update OpenTelemetry dependencies to 0.17.0
+- Add support for `ottrace` propagator
 
 ## 0.16.1
 - Enable gzip compression by default